Use "shadow" copy of delay buffer for specialized implementation
By introducing a second "shadow" copy of the delay buffer right after
the original one in memory, we can always find an unbroken memory
segment that contains all the samples we are interested in. This allows
us to use a single inner loop instead of having to break it up into two
parts, at the cost of some extra memory.